Reorganized cooltools into include src lib and bin folders, updated Makefile
